requirements.txt
pydantic==2.7.0
python-dotenv==1.0.0
PyYAML==6.0.1
pandas==2.3.1
openpyxl==3.1.5
pymupdf==1.26.3
playwright==1.55.0
streamlit

